Fix Windows build: split flock into platform-specific files
syscall.Flock is not available on Windows. Split the implementation
into flock_unix.go (build tag !windows) with the real flock logic
and flock_windows.go that returns an error. The interface, constructor,
and helper functions stay in pushlock.go.

Windows builds compile but push lock is unsupported at runtime.
